refactor(server-remote-standalone): inline types, middleware, and cors into server.ts and auth.ts
